+++ CHANGELOG.md
@@ -4,9 +4,12 @@
 This release extends support to OneView Rest API version 1800 (OneView v5.30) for the hardware variants C7000 and Synergy to the following features:
 
 #### Features supported
-- Ethernet Network
 - Enclosure
 - Enclosure Group
+- Ethernet Network
+- FC Network
+- FCOE Network
+- Scope
 
 ## v5.13.0